set up base activity
