+++ frontend/app/src/main/AndroidManifest.xml
@@ -16,7 +16,9 @@
         android:theme="@style/Theme.MyApplication"
         tools:remove="android:appComponentFactory"
         tools:targetApi="31">
-
+        <activity
+            android:name=".BaseActivity"
+            android:exported="false" />
         <activity
             android:name=".LoginActivity"
             android:exported="true">
@@ -35,7 +37,7 @@
             android:name=".SubscribedArticlesActivity"
             android:exported="false"
             android:label="@string/title_activity_subscribed"
-            android:theme="@style/Theme.MyApplication.NoActionBar"/>
+            android:theme="@style/Theme.MyApplication.NoActionBar" />
         <activity
             android:name=".ForumActivity"
             android:exported="false"
@@ -48,7 +50,7 @@
             android:theme="@style/Theme.MyApplication.NoActionBar" />
         <activity
             android:name=".MainActivity"
-            android:exported="false"/>
+            android:exported="false" />
         <activity
             android:name=".HistoryActivity"
             android:exported="false"
